move global Qt app setup to main
