+++ src/app/main.cpp
@@ -20,16 +20,17 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <QTextCodec>
 
 #include <csignal>
 
-#include "runtime.h"
+#include <QTextCodec>
+#include <QGuiApplication>
+#include <QStyleHints>
+#include <QQuickWindow>
+
 #include "log.h"
 
 #include "app.h"
-
-#include "framework/global/globalmodule.h"
 
 #include "muse_framework_config.h"
 
@@ -245,6 +246,11 @@ static void crashCallback(int signum)
 
 #endif
 
+static void app_init_qrc()
+{
+    Q_INIT_RESOURCE(app);
+}
+
 int main(int argc, char** argv)
 {
 #ifndef MUSE_MODULE_DIAGNOSTICS_CRASHPAD_CLIENT
@@ -253,9 +259,69 @@ int main(int argc, char** argv)
     signal(SIGFPE, crashCallback);
 #endif
 
+    // ====================================================
+    // Setup global Qt application variables
+    // ====================================================
+
     // Force the 8-bit text encoding to UTF-8. This is the default encoding on all supported platforms except for MSVC under Windows, which
     // would otherwise default to the local ANSI code page and cause corruption of any non-ANSI Unicode characters in command-line arguments.
     QTextCodec::setCodecForLocale(QTextCodec::codecForName("UTF-8"));
+
+    app_init_qrc();
+
+    qputenv("QT_STYLE_OVERRIDE", "Fusion");
+    qputenv("QML_DISABLE_DISK_CACHE", "true");
+
+#ifdef Q_OS_LINUX
+    if (qEnvironmentVariable("QT_QPA_PLATFORM") != "offscreen") {
+        qputenv("QT_QPA_PLATFORMTHEME", "gtk3");
+    }
+#endif
+
+#ifdef Q_OS_WIN
+    // NOTE: There are some problems with rendering the application window on some integrated graphics processors
+    //       see https://github.com/musescore/MuseScore/issues/8270
+    QCoreApplication::setAttribute(Qt::AA_UseOpenGLES);
+
+    if (!qEnvironmentVariableIsSet("QT_OPENGL_BUGLIST")) {
+        qputenv("QT_OPENGL_BUGLIST", ":/resources/win_opengl_buglist.json");
+    }
+#endif
+
+#ifdef MU_QT5_COMPAT
+    QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
+#endif
+
+//! NOTE: For unknown reasons, Linux scaling for 1 is defined as 1.003 in fractional scaling.
+//!       Because of this, some elements are drawn with a shift on the score.
+//!       Let's make a Linux hack and round values above 0.75(see RoundPreferFloor)
+#ifdef Q_OS_LINUX
+    QGuiApplication::setHighDpiScaleFactorRoundingPolicy(Qt::HighDpiScaleFactorRoundingPolicy::RoundPreferFloor);
+#elif defined(Q_OS_WIN)
+    QGuiApplication::setHighDpiScaleFactorRoundingPolicy(Qt::HighDpiScaleFactorRoundingPolicy::PassThrough);
+#endif
+
+    QGuiApplication::styleHints()->setMousePressAndHoldInterval(250);
+
+#ifndef MU_QT5_COMPAT
+    // Necessary for QQuickWidget, but potentially suboptimal for performance.
+    // Remove as soon as possible.
+    QQuickWindow::setGraphicsApi(QSGRendererInterface::OpenGL);
+#endif
+
+    QCoreApplication::setApplicationName(MUSE_APP_TITLE);
+    QCoreApplication::setOrganizationName("MuseScore");
+    QCoreApplication::setOrganizationDomain("musescore.org");
+    QCoreApplication::setApplicationVersion(MUSE_APP_VERSION);
+
+#if !defined(Q_OS_WIN) && !defined(Q_OS_DARWIN) && !defined(Q_OS_WASM)
+    // Any OS that uses Freedesktop.org Desktop Entry Specification (e.g. Linux, BSD)
+#ifndef MUSE_APP_INSTALL_SUFFIX
+#define MUSE_APP_INSTALL_SUFFIX ""
+#endif
+    QGuiApplication::setDesktopFileName("org.musescore.MuseScore" + QString(MUSE_APP_INSTALL_SUFFIX) + ".desktop");
+#endif
+
 
     mu::app::App app;
